src/backend/pyproject.toml
[tool.poetry]
name = "cesi-2022-cube"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.70.0"}
sqlmodel = "^0.0.4"
alembic = "^1.7.5"
psycopg2-binary = "^2.9.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
